+++ src/main/java/de/jost_net/JVerein/io/AuswertungMitgliedschaftsjubilarePDF.java
@@ -29,10 +29,11 @@ import de.jost_net.JVerein.io.Adressbuch.Adressaufbereitung;
 import de.jost_net.JVerein.keys.VorlageTyp;
 import de.jost_net.JVerein.rmi.Mitglied;
 import de.jost_net.JVerein.util.VorlageUtil;
-import de.willuhn.logging.Logger;
 
-public class AltersjubilaeumsExportPDF extends AltersjubilaeumsExport
+public class AuswertungMitgliedschaftsjubilarePDF
+    extends AuswertungMitgliedschaftsjubilareAbstract
 {
+
   private FileOutputStream fos;
 
   private Reporter reporter;
@@ -42,7 +43,7 @@ public class AltersjubilaeumsExportPDF extends AltersjubilaeumsExport
   @Override
   public String getName()
   {
-    return "Altersjubilare PDF-Export";
+    return "Mitgliedschaftsjubilare PDF-Export";
   }
 
   @Override
@@ -54,10 +55,11 @@ public class AltersjubilaeumsExportPDF extends AltersjubilaeumsExport
     }
     IOFormat f = new IOFormat()
     {
+
       @Override
       public String getName()
       {
-        return AltersjubilaeumsExportPDF.this.getName();
+        return AuswertungMitgliedschaftsjubilarePDF.this.getName();
       }
 
       /**
@@ -75,15 +77,15 @@ public class AltersjubilaeumsExportPDF extends AltersjubilaeumsExport
   @Override
   public String getDateiname(Object object)
   {
-    return VorlageUtil.getName(VorlageTyp.AUSWERTUNG_ALTERSJUBILARE_DATEINAME,
-        object) + ".pdf";
+    return VorlageUtil.getName(
+        VorlageTyp.AUSWERTUNG_MITGLIEDSCHAFTSJUBILARE_DATEINAME, object)
+        + ".pdf";
   }
 
   @Override
   protected void open() throws DocumentException, IOException
   {
     fos = new FileOutputStream(file);
-    Logger.debug(String.format("Altersjubilare, Jahr=%d", jahr));
     reporter = new Reporter(fos, params.getTitle(), params.getSubtitle(),
         params.getLinks(), params.getRechts(), params.getOben(),
         params.getUnten(), false, params.getVordergrund(),
@@ -94,12 +96,11 @@ public class AltersjubilaeumsExportPDF extends AltersjubilaeumsExport
   @Override
   protected void startJahrgang(int jahrgang) throws DocumentException
   {
-    Logger.debug(String.format("Altersjubiläum, Jahrgang=%d", jahrgang));
     Paragraph pHeader = new Paragraph(
-        "\n" + String.format("%d. Geburtstag", jahrgang),
+        "\n" + String.format("%d-jähriges Jubiläum", jahrgang),
         Reporter.getFreeSans(11));
     reporter.add(pHeader);
-    reporter.addHeaderColumn("Geburtsdatum", Element.ALIGN_CENTER, 50,
+    reporter.addHeaderColumn("Eintrittsdatum", Element.ALIGN_CENTER, 50,
         params.getColorHeader(), params.getFontHeader());
     reporter.addHeaderColumn("Name, Vorname", Element.ALIGN_CENTER, 100,
         params.getColorHeader(), params.getFontHeader());
@@ -128,7 +129,7 @@ public class AltersjubilaeumsExportPDF extends AltersjubilaeumsExport
   @Override
   protected void add(Mitglied m) throws RemoteException
   {
-    reporter.addColumn(m.getGeburtsdatum(), Element.ALIGN_LEFT,
+    reporter.addColumn(m.getEintritt(), Element.ALIGN_LEFT,
         params.getFontNormal());
     reporter.addColumn(Adressaufbereitung.getNameVorname(m), Element.ALIGN_LEFT,
         params.getFontNormal());
